# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hover API
  version: 0.4.1
  description: |
    Lean OpenAPI coverage for Hover's current backend API surface, optimised for
    Postman import, backend testing, and API exploration.

    The spec documents implemented behaviour from the Go handlers rather than
    idealised REST shapes.
servers:
  - url: https://hover.app.goodnative.co
    description: Production application
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase bearer token sent in the `Authorization` header.
...
  schemas:
    ErrorResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: Invalid JSON request body
        code:
          type: string
          example: BAD_REQUEST
        request_id:
          type: string
          example: req_123abc
...
    SuccessEnvelope:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: success
        data:
          nullable: true
        message:
          type: string
        request_id:
          type: string
          example: req_123abc
...
    Scheduler:
      type: object
      required:
        - id
        - domain
        - schedule_interval_hours
        - next_run_at
        - is_enabled
        - concurrency
        - find_links
        - max_pages
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        domain:
          type: string
        schedule_interval_hours:
          type: integer
          enum: [6, 12, 24, 48]
        next_run_at:
          type: string
          format: date-time
        is_enabled:
          type: boolean
        concurrency:
          type: integer
        find_links:
          type: boolean
        max_pages:
          type: integer
        include_paths:
          type: array
          items:
            type: string
        exclude_paths:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SchedulerRequest:
      type: object
      properties:
        domain:
          type: string
          description: Required on create, ignored on update.
        schedule_interval_hours:
          type: integer
          enum: [6, 12, 24, 48]
        concurrency:
          type: integer
        find_links:
          type: boolean
        max_pages:
          type: integer
        include_paths:
          type: array
          items:
            type: string
        exclude_paths:
          type: array
          items:
            type: string
        is_enabled:
          type: boolean
        expected_is_enabled:
          type: boolean
          description: Optional optimistic concurrency hint on update.
...
    WebflowSiteSetting:
      type: object
      required:
        [webflow_site_id, site_name, primary_domain, auto_publish_enabled]
      properties:
        webflow_site_id:
          type: string
        site_name:
          type: string
        primary_domain:
          type: string
        last_published:
          type: string
        schedule_interval_hours:
          type: integer
          nullable: true
        auto_publish_enabled:
          type: boolean
        scheduler_id:
          type: string
          format: uuid
          nullable: true
...
    UpdateWebflowSiteScheduleRequest:
      type: object
      required: [connection_id]
      properties:
        connection_id:
          type: string
          format: uuid
        schedule_interval_hours:
          type: integer
          enum: [6, 12, 24, 48]
          nullable: true
          description: Send `null` to remove the schedule.
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorised:
      description: Authentication failed or token missing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /v1/schedulers:
    get:
      tags: [Jobs]
      summary: List schedulers for the active organisation
      responses:
        "200":
          description: Scheduler list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Scheduler"
        "401":
          $ref: "#/components/responses/Unauthorised"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Jobs]
      summary: Create a scheduler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchedulerRequest"
            example:
              domain: example.com
              schedule_interval_hours: 24
              concurrency: 20
              find_links: true
              max_pages: 0
              include_paths: [/blog/*]
              exclude_paths: [/admin/*]
      responses:
        "201":
          description: Scheduler created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Scheduler"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorised"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/schedulers/{id}:
    get:
      tags: [Jobs]
      summary: Get a scheduler
      parameters:
        - name: id
          in: path
          required: true
          description: Scheduler UUID.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Scheduler details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Scheduler"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [Jobs]
      summary: Update a scheduler
      parameters:
        - name: id
          in: path
          required: true
          description: Scheduler UUID.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchedulerRequest"
      responses:
        "200":
          description: Scheduler updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Scheduler"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Scheduler state conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Jobs]
      summary: Delete a scheduler
      parameters:
        - name: id
          in: path
          required: true
          description: Scheduler UUID.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Scheduler deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [id, status]
                        properties:
                          id:
                            type: string
                            format: uuid
                          status:
                            type: string
                            example: deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/integrations/webflow/sites/{site_id}/schedule:
    put:
      tags: [Integrations]
      summary: Create, update, or remove a Webflow site schedule
      parameters:
        - name: site_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWebflowSiteScheduleRequest"
      responses:
        "200":
          description: Site schedule updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/WebflowSiteSetting"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...